[Monopoly - Jail]
Implemented jail for the game. The implementation thus far doesn't include drawing cards or rolling three doubles in a turn resulting in jail. The framework is built for these scenarios however.
Also found a bug where the setting of location wasn't accurately placing the player on the properties, fixed this and added some tests for this scenario.
Added some tests for the jail scenarios