[Chance and Community Cards]
Implemented chance and community chest cards within the game. This includes the specific functionality around "Get out of Jail free" cards and "Go to Jail" cards. Implemented the setting up of cards so it is done through the FileReader similar to the setting up of properties.
Added tests to cover the card functionality. (coverage at 80%)