[Saving/Loading Banker]
When loading properties the banker needs to be set as the owner for properties that they were the owner of originally, this wasn't happening properly so added a fix and extended the test to cover this.
Also made it so we save and load the actual banker object being used in the game, this isn't part of the board so wasn't being persisted when loading a new game. Added a test for this